refactor: index files added for components
