+++ src/App.jsx
@@ -1,25 +1,23 @@
-import Hero from './Components/Hero/Hero'
-import './App.css'
-import Populer from './Components/Populer/Populer'
-import Teklifler from './Components/Teklifler/Teklifler'
-import Newsletter from './Components/Newsletter/Newsletter'
-import YeniUrunler from './Components/YeniUrunler/YeniUrunler'
+import Hero from "./Components/Hero";
+import "./App.css";
+import Populer from "./Components/Populer";
+import Teklifler from "./Components/Teklifler";
+import Newsletter from "./Components/Newsletter";
+import YeniUrunler from "./Components/YeniUrunler";
 
 function App() {
-
   return (
     <>
-    <Hero></Hero>
-    <Populer></Populer>
-    <Teklifler></Teklifler>
-    <YeniUrunler></YeniUrunler>
-    <Newsletter></Newsletter>
+      <Hero></Hero>
+      <Populer></Populer>
+      <Teklifler></Teklifler>
+      <YeniUrunler></YeniUrunler>
+      <Newsletter></Newsletter>
     </>
-  )
+  );
 }
 
-export default App
-
+export default App;
 
 /*
