+++ src/Components/UserManagement/index.jsx
@@ -1,0 +1,1 @@
+export { default } from "./UserManagement";
