+++ src/Pages/Market.jsx
@@ -1,19 +1,17 @@
-import Newsletter from "../Components/Newsletter/Newsletter";
-import Populer from "../Components/Populer/Populer";
-import Teklifler from "../Components/Teklifler/Teklifler";
-import YeniUrunler from "../Components/YeniUrunler/YeniUrunler";
+import Newsletter from "../Components/Newsletter";
+import Populer from "../Components/Populer";
+import Teklifler from "../Components/Teklifler";
+import YeniUrunler from "../Components/YeniUrunler";
 
-function Market(){
-
-    return(
-        <>
-        <Populer></Populer>
-        <Teklifler></Teklifler>
-        <YeniUrunler></YeniUrunler>
-        <Newsletter></Newsletter>
-        </>
-    )
-
+function Market() {
+  return (
+    <>
+      <Populer></Populer>
+      <Teklifler></Teklifler>
+      <YeniUrunler></YeniUrunler>
+      <Newsletter></Newsletter>
+    </>
+  );
 }
 
 export default Market;
