+++ src/api/axios.js
@@ -1,0 +1,8 @@
+// src/api/axios.js
+import axios from "axios";
+
+const instance = axios.create({
+  baseURL: "http://localhost:3000/api",
+});
+
+export default instance;
